block_allocator: add cc
